Change Server.new and the class
Server.new now only needs a host and accepts name, start_date and
players optionally.
The class Server no longer stores login data and server data as an
object to which it delegates methods. Now the data from these two
sources is saved to regular instance variables.
Login data is loaded on initialization for obvious reasons and the
server data is load upon request to any of the methods :world_id,
:version, :speed or restart_date by the private method :server_data.

Signed-off-by: João Soares <jsoaresgeral@gmail.com>